backend/src/config.py
import os
from dotenv import load_dotenv

# Load .env file
load_dotenv()

class Config:
    DEBUG=False
    TESTING = False
    SECRET_KEY = os.getenv('SECRET_KEY')
    JWT_SECRET_KEY = os.getenv('JWT_SECRET_KEY')
    SQLALCHEMY_DATABASE_URI = os.getenv('DATABASE_URI')
    SQLALCHEMY_TRACK_MODIFICATIONS = False
    